refactor: extract creatives controller concerns
Split CreativesController into focused concerns to improve maintainability:
- SlideViewable: slide_view action and build_slide_ids helper
- Exportable: export_markdown action
- TreeManageable: children, reorder, link_drop, append_*, unconvert actions
- Shareable: request_permission action

Reduced controller from 503 to 303 lines while maintaining all functionality.
All tests passing.
